edited 4.1.3 by mentor's advices
Fit vertically into the window and make it "rubbery"

.btn__reset is simply moved to normalize and specified for all buttons, so as not to prescribe this class for each button.

.btn__search, btn__feedback, and so on are modifiers for the btn Block. It's worth choosing a more specific name instead of btn

Instead of putting individual parts in a .container, you can simply wrap everything aside and set the margins on the sides if they are the same everywhere.

Try not to use increased specificity unnecessarily

Do not use negative external margins. Especially for Blocks

It is not necessary to wrap everything in nav and ul. In this case, nav and ul can only be left at .main__nav

Since this is part of the overall layout, main will be used here.